requirements.txt
asgiref==3.4.1
Django==3.2.6
django-ckeditor==6.1.0
django-js-asset==1.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Pillow==8.3.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
typing==3.7.4.3
